refactor: split pagewrapper components
