+++ apps/journeys-admin/src/components/PageWrapper/NavigationDrawer/DrawerContent/UserMenu/index.ts
@@ -1,0 +1,1 @@
+export { UserMenu } from './UserMenu'
