+++ apps/journeys-admin/src/components/PageWrapper/NavigationDrawer/DrawerContent/index.ts
@@ -1,0 +1,1 @@
+export { DrawerContent } from './DrawerContent'
