+++ apps/journeys-admin/src/components/PageWrapper/NavigationDrawer/index.ts
@@ -1,0 +1,1 @@
+export { NavigationDrawer } from './NavigationDrawer'
